requirements.txt
playwright==1.32.1
pytest-playwright==0.3.2
pytest==7.2.0
requests==2.26.0
pytest-asyncio==0.20.2
allure-pytest==2.9.45
pytest-rerunfailures==10.2
virtualenv==20.16.6
faker==15.3.2
python-dotenv==0.21.1
pyyaml==6.0
pydantic==1.10.4